feat(FogNode): Added initial implementation of the FogNode module
For the time being it is isolated from the rest of the topology
for testing purposes.

A new scenario is included for the interaction between a single
datacentre and a fog node. Right now it only requests a vm, that
will in the future be used for the execution of an app simulating
a workload.

Probably it would be useful to create a message that encapsulates
both the service request and the rental of the resources.