feat(FogNode): Implemented a simple IO request simulation
Currently this node only supports an IO request (modeled by SM_FogIO)
and will always return OK (it's represented as 0 in getResult()).

Next action should be to implement this node in conjuction with edge
devices (IotController) and research the possibility of recording the
network statistics (used bandwidth in a DatarateChannel).