feat(FogNode,IotController,IotSensor): Implemented base IoT prototypes
Now there are two networks inside the ini file so that both
scenarios (Fog to DC and Sensor-Controller-Fog) can be run in the same space

The current model is the next one (in the future it could be parametrized):

- The IotSensor sends periodically data (what's referred as "Edge pings")

- After 10 "Edge pings" from an IotSensor, the IotController sends an write request
to the Fog (simulating data aggregation).

- After 10 "IO requests" the Fog initializes the process of requesting an VM and a
service from the DC. It should be noted that this scenario is currently unstable
(will be fixed shortly).

Also there's the prototype of the request in the class SM_FogIO, which in conjuction
with an datarate channel, it can simulate the transmission time based on it's size.
It should be noted that for simulating the real thing, there needs to be the propagation delay.