feat(MessageAdapter,NetworkAdapter): Integrated successfully SIMCAN with INET
With the definition of the wrapper Chunk (INET_AppMessage), two SIMCAN nodes
can communicate through the network relying on their respective NetworkAdapters.

In the future it will be necessary to introduce a DNS like system in order to be
able to address different nodes interconnected through the INET network. An idea
that might be able to work is to dump the configurator file with all the IPs and
create a module that parses that file (getEnvir()->getXMLDocument())

Also it's important to reflect on how to install and link the INET project to SIMCAN
in the documentation.

I have also detected bugs in the dup() method of most of the messages defined.
I will do a refactor that corrects them, following the recommendations of the SimulationManual.