feat(DNS_Resolver,RequestMultiplexer): Implemented first version of the DNS_Resolver
Currently the DNS request structure support more heavy duty requests than the Resolver
currently issues (it requests one resolution at a time). In the future it could be reimplemented
to take more advantage of it's cache and pending requests. There's also the addition of the
RequestMultiplexer, an module which helps coordinate the services inside the NetworkAdapter.

There are also minor fixes (improved testcases) and GUI updates for better simulation visualization.