test(DNS_Service,DNS_Resolver): Added extra testing and showcased tested functionality
Added the test DNS_Service_Resolution.test which tests in a simple level (to be expanded)
the cycle for a NS resolution (in this case the kind of a DataCentre NS). Showcased also
this function in action under INETcomm where the prototype FogNode resolves first the DataCenters
IP and then sends the subsequent requests.