feat(SM_REST_API): Defined the basic message for API calls
Included the base class declaration for API calls. The final class should include also cookies
and a mechanism for determining and auto updating the payload size. Also improved the EdgeHypervisor.
On another note included the first label system for syscalls. It should help for retracking the original
requester (pid and vmId).