feat(SimSchemaMySQL): Defined interface and MySQL implementation of SimSchema
The mission of SimSchema is to abstract the experiment configuration from the models.
The implementation based on the DB MySQL (SimSchemaMySQL) is included with testing
to ensure the correct behavior. In order to use it one must create a MySQL database
and pass the right connection string to the module. Note that the password is read
from the environement variable S2F_DB_PASSWORD for safety.