fix: General fixes and refactoring
The code readability has improved significantly. This is due to assigning to DataManager the
responsibility of retrieving and interpreting the data (talking mostly with SimSchema). There
were also major refactors improving code reusability and modern C++ features. In a not so distant
future I will have to tackle the warnings of methods not marked with the override keyword.