feat(ApplicationBuilder): Added the module responsible for app instantiation
Again, this shift of responsibility simplifies greatly the step of application
creation and instantation. For the moment the ApplicationBuilder class acts as an
interface but in the future it will probably contain the common steps.

I also added the necessary configurations in order to run the demo network with
all of the recent changes.