feat(OsCore): Common operating system utilities that help hypervisors
This relieves the hypervisor module the responsibility of handling and processing
system calls, so it focuses in topological reasoning and message routing. Also
added a general refactor for ApplicationBuilder so it contains the common steps
for DC and Edge environements.

In another note I refactored the SM_UserAPP message in order to be more efficient
and less convoluted in it's own interface.