feat(Hypervisor, ResourceManager): Redefinition of modules
Currently it's in early development stages, but the idea is to detach the DataCentreManager
module from the details inside a Blade (Hypervisor + HardwareManager). In order to do this
there must be the correct interfaces and the neccesary networking to communicate the blades
with the manager. This is also a great step in allowing the outer communication of the
applications inside a DataCenter.