feat(Hypervisor): Hypervisor super general class
Implemented VmControlBlocks and generic class ControlTable<T> that will allow
easier management of both VMs and Apps. As a result, the implementation is becoming
more general and subclasses mostly need to fill in a few details. OsCore now interacts
with the superclass instead (before it only did it with EdgeHypervisor).