feat(DcResourceManager): Resource Management utilities for Data Centers
This module follows the line of reducing the responsibilities of the DcManagers,
handling all the inner logic related with resource management. The most important
feature will be the support for "transactional deployments" which will allow for a more
efficient rollback in case the deployment (SM_UserVM) is just not viable.

Other minor changes are minor fixes and also displacing the logic of the timing of
cpu usage from the DcManager to the HardwareManager.