feat(AppProxy, RestMsg): Implemented App Proxy with Restful Messages
Major redesign over the approach of the Proxy implementation.
The core ideas are to use the "in band" approach so it's easier
to manage the connections, sessions and redirections.

There's room for improvement, yet this approach shows potential
and even hint's for a completely detached version of this module.

Checkout the demo in showcases/proxy/ folder. Next is an example
involving the DNS resolution and protocol.