feat(management): Revamped the whole management package
The most important addition is the module NodePool. As of today there
will only be one. However, this opens to a future oportunity to deploy
different NodePools that could aggregate nodes based on characteristics,
pricing, priority, etc. Another improvement is the drastic reduction in
complexity in the logic of the ResourceManager/SelectionStragegy.