feat(s2f,run-sim): Created universal scripts for running the sim
The s2f script, included under the bin/ folder of the project, allows
to run the simulator in all form factors (as a dynamic lib or executable).

The run-sim takes it a step further. When the user is on a simulation folder
(under the simulations/ folder) it will launch the corresponding experiment,
giving the user the choice to run any of the available configurations.