fix: incorrect connection finish on server-side
When two peers A and B both see each other as potential candidates
(either through dns discovery or via acquired addr messages), whoever
received a connection request from the other would see them as a
'connected' peer and close the connection when it should not.

The current work-around for this (knowing no two connections can occur
at the same time because of the event scheduling) is for the receiver (B)
node to update peer information for a candidate (A) that connected to
it.